server/requirements.txt
asgiref==3.11.0
attrs==25.4.0
cattrs==25.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
flatbuffers==25.9.23
h11==0.16.0
idna==3.11
jh2==5.0.10
niquests==3.15.2
numpy==2.4.0
openmeteo_requests==1.7.4
openmeteo_sdk==1.23.0
pandas==2.3.3
platformdirs==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qh3==1.5.6
requests==2.32.5
requests-cache==1.2.1
retry-requests==2.0.0
six==1.17.0
sqlparse==0.5.4
stripe==14.1.0
typing_extensions==4.15.0
tzdata==2025.3
url-normalize==2.2.1
urllib3==2.6.2
urllib3-future==2.15.900
wassima==2.0.3